+++ src/dev/js/main.js
@@ -2,7 +2,7 @@
 jQuery.fn.exists = function () {
     return $(this).length;
 }
-
+$('.fancybox').fancybox();
 $(() => {
 
     let parentEl = '';
@@ -835,43 +835,45 @@ $(() => {
         }
     }
 
-    var btn = document.getElementById("theme-button");
-    var link = document.getElementById("theme-link");
-    var themeDefault = "";
+    if ($('#theme-button').exists()) {
+        var btn = document.getElementById("theme-button");
+        var link = document.getElementById("theme-link");
+        var themeDefault = "";
 
-    btn.addEventListener("click", function () {
-        let theme = ChangeTheme();
-        sessionStorage.setItem('theme', theme)
-    });
+        btn.addEventListener("click", function () {
+            let theme = ChangeTheme();
+            sessionStorage.setItem('theme', theme)
+        });
 
-    if (sessionStorage.getItem('theme') !== null) {
-        $('#theme-link').attr('href', sessionStorage.getItem('theme'));
-        if (sessionStorage.getItem('theme') == themeDefault) {
+        if (sessionStorage.getItem('theme') !== null) {
+            $('#theme-link').attr('href', sessionStorage.getItem('theme'));
+            if (sessionStorage.getItem('theme') == themeDefault) {
+                $('.header .checkbox input').prop('checked', true);
+            } else {
+                $('.header .checkbox input').prop('checked', false);
+            }
+        } else {
+            $('#theme-link').attr('href', themeDefault);
             $('.header .checkbox input').prop('checked', true);
-        } else {
-            $('.header .checkbox input').prop('checked', false);
-        }
-    } else {
-        $('#theme-link').attr('href', themeDefault);
-        $('.header .checkbox input').prop('checked', true);
-    }
-
-    function ChangeTheme() {
-        let lightTheme = themeDefault;
-        let darkTheme = "css/dark.css";
-
-        var currTheme = link.getAttribute("href");
-        var theme = "";
-
-        if (currTheme == lightTheme) {
-            currTheme = darkTheme;
-            theme = "dark";
-        } else {
-            currTheme = lightTheme;
-            theme = "light";
         }
 
-        link.setAttribute("href", currTheme);
-        return currTheme;
+        function ChangeTheme() {
+            let lightTheme = themeDefault;
+            let darkTheme = "/local/templates/magmetall/css/dark.css ";
+
+            var currTheme = link.getAttribute("href");
+            var theme = "";
+
+            if (currTheme == lightTheme) {
+                currTheme = darkTheme;
+                theme = "dark";
+            } else {
+                currTheme = lightTheme;
+                theme = "light";
+            }
+
+            link.setAttribute("href", currTheme);
+            return currTheme;
+        }
     }
 });
